Revert "Change Data for chart & export and export data processing (#383)"
This reverts commit b5ba31e5f091bf742f725e0057f3a67f02145009.
