+++ src/app/Main/Body/InfoPage/InfoPage-FiltersContent.component.tsx
@@ -104,7 +104,6 @@ export function InfoPageFiltersContentComponent(props: {
                         describedFiltersContent
                     )}
                     defaultExpanded={false}
-                    centerContent={false}
                     key={`accordion_${describedFilter}`}
                 />
             );
@@ -133,7 +132,6 @@ export function InfoPageFiltersContentComponent(props: {
                     title={t(`Filters.resistance.Name`)}
                     content={resistanceContent}
                     defaultExpanded={false}
-                    centerContent={false}
                     key="accordion_resistance"
                 />
             );
@@ -148,7 +146,6 @@ export function InfoPageFiltersContentComponent(props: {
                         </div>
                     }
                     defaultExpanded={false}
-                    centerContent={false}
                     key="accordion_samplingContext"
                 />
             );
@@ -158,7 +155,6 @@ export function InfoPageFiltersContentComponent(props: {
                     title={t(`Filters.${describedFilter}.Name`)}
                     content={t(`Filters.${describedFilter}.Description`)}
                     defaultExpanded={false}
-                    centerContent={false}
                     key={`accordion_${describedFilter}`}
                 />
             );
