+++ src/app/Main/Body/InfoPage/InfoPage.component.tsx
@@ -117,7 +117,6 @@ export function InfoPageComponent(props: {
                         </div>
                     }
                     defaultExpanded={false}
-                    centerContent={false}
                 />
                 <p css={subHeadingStyle} id="filter">
                     {filtersChapterHeading}
@@ -138,7 +137,6 @@ export function InfoPageComponent(props: {
                             </div>
                         }
                         defaultExpanded={false}
-                        centerContent={false}
                     />
                     <AccordionComponent
                         title={t("Methods.Amrs.Name")}
@@ -150,7 +148,6 @@ export function InfoPageComponent(props: {
                             />
                         }
                         defaultExpanded={false}
-                        centerContent={false}
                     />
                 </div>
             </div>
