+++ src/app/Main/Body/QueryPage/QueryPageContent/NumberOfIsolates/QueryPage-NrOfIsolates.component.tsx
@@ -44,7 +44,6 @@ export function QueryPageNrOfIsolatesComponent(props: {
                 </table>
             }
             defaultExpanded
-            centerContent={false}
         />
     );
 }
