+++ src/app/Main/Body/QueryPage/QueryPageContent/Parameter/QueryPage-ParameterContent.component.tsx
@@ -48,7 +48,6 @@ export function QueryPageParameterContentComponent(props: {
                 </div>
             }
             defaultExpanded
-            centerContent={false}
         />
     );
 }
