+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/QueryPageContent-TableResultsLayout.component.tsx
@@ -1,13 +1,20 @@
-import React, { useState } from "react";
+/** @jsx jsx */
+import { css, jsx } from "@emotion/core";
+import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import _ from "lodash";
+import { AccordionComponent } from "../../../../../../Shared/Accordion.component";
 import { ResultsTableOptionsComponent } from "./TableResults-Options.component";
 import { ExplanationTextComponent } from "../../../../../../Shared/ExplanationText.component";
 import { DataInterface } from "../../../../../../Shared/Context/DataContext";
-import { TableContainer } from "./Table/Table-Container.component";
+import { TableResultsTableComponent } from "./TableResults-Table.component";
 import { SumOptions } from "./TableResults.model";
-import { LoadingProcessComponent } from "../../../../../../Shared/LoadingProcess.component";
-import { AccordionComponent } from "../../../../../../Shared/Accordion.component";
+
+const dataStyle = css`
+    max-width: fit-content;
+    margin: auto;
+    box-sizing: inherit;
+`;
 
 export interface TableResultsProps {
     tableIsLoading: boolean;
@@ -21,7 +28,9 @@ export interface TableResultsProps {
  * @param props - info about isCol/isRow and columnAttributes
  * @returns {JSX.Element} - accordion with the result table
  */
-export function TableResultsContainer(props: TableResultsProps): JSX.Element {
+export function QueryPageContentTableResultsLayoutComponent(
+    props: TableResultsProps
+): JSX.Element {
     const [sumOptions, setSumOptions] = useState<SumOptions>({
         showRowSum: true,
         showColSum: true,
@@ -46,26 +55,30 @@ export function TableResultsContainer(props: TableResultsProps): JSX.Element {
         ? ""
         : t(`Filters.${tableColAttribute}`);
 
-    let tableAccordionContent = <ExplanationTextComponent />;
-    const isTable: boolean =
-        props.tableData.row !== "" || props.tableData.column !== "";
+    let tableAccordionContent = (
+        <div css={dataStyle}>
+            <ExplanationTextComponent />
+        </div>
+    );
+    const isTable: boolean = props.tableData.row !== "" || props.tableData.column !== ""
 
-    let tableData = props.tableData.statTableDataWithSumsAbs;
+    let tableData = props.tableData.statisticDataAbsolute;
     if (props.tableData.option === "relative") {
-        tableData = props.tableData.statTableDataWithSumsRel;
+        tableData = props.tableData.statisticDataRelative;
     }
 
-    if (isTable && !props.tableIsLoading) {
+    if (isTable) {
         tableAccordionContent = (
-            <div>
+            <div css={dataStyle}>
                 <ResultsTableOptionsComponent
                     sumOptions={sumOptions}
                     tableOption={props.tableData.option}
                     onDisplayOptionsChange={handleChangeDisplayOptions}
                     onChangeSumOptions={handleSumOptionsChange}
                 />
-                <TableContainer
+                <TableResultsTableComponent
                     sumOptions={sumOptions}
+                    isLoading={props.tableIsLoading}
                     colMainHeader={colMainHeader}
                     rowMainHeader={rowMainHeader}
                     tableData={tableData}
@@ -78,26 +91,11 @@ export function TableResultsContainer(props: TableResultsProps): JSX.Element {
         );
     }
 
-    if (isTable && props.tableIsLoading) {
-        tableAccordionContent = (
-            <div>
-                <ResultsTableOptionsComponent
-                    sumOptions={sumOptions}
-                    tableOption={props.tableData.option}
-                    onDisplayOptionsChange={handleChangeDisplayOptions}
-                    onChangeSumOptions={handleSumOptionsChange}
-                />
-                <LoadingProcessComponent />
-            </div>
-        );
-    }
-
     return (
         <AccordionComponent
             title={accordionHeader}
             content={tableAccordionContent}
             defaultExpanded
-            centerContent
         />
     );
 }
