+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/TableContent-Header.component.tsx
@@ -3,15 +3,15 @@ import { css, jsx } from "@emotion/core";
 import { withStyles, createStyles } from "@material-ui/core/styles";
 import TableCell from "@material-ui/core/TableCell";
 import { useTranslation } from "react-i18next";
-import { DisplayOptionType } from "../../../../../../../Shared/Context/DataContext";
-import { onBackgroundColor } from "../../../../../../../Shared/Style/Style-MainTheme.component";
+import { DisplayOptionType } from "../../../../../../Shared/Context/DataContext";
+import { onBackgroundColor } from "../../../../../../Shared/Style/Style-MainTheme.component";
 import {
     highlightedTableBorder,
     fixedCellSize,
     defaultTableBorder,
     sumRowColBackgroundColor,
-} from "../ResultsTable.style";
-import { getMicroorganismLabelService } from "../../../../Services/getMicroorganismLabel";
+} from "./ResultsTable.style";
+import { getMicroorganismLabelService } from "../../../Services/getMicroorganismLabel";
 
 const blankCellStyle = css`
     box-sizing: border-box;
@@ -53,7 +53,7 @@ const StyledTableCell = withStyles(() =>
  * @param headerValues - object with two booleans, true if row/column is selected
  * @returns {JSX.Element[]} - list of table cell components
  */
-export function createTableHeaderService(props: {
+export function TableContentHeaderComponent(props: {
     showRowSum: boolean;
     headerValues: string[];
     colAttribute: string;
