+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/TableContent-RowWithColSums.component.tsx
@@ -1,13 +1,12 @@
 /** @jsx jsx */
 import { css, jsx, SerializedStyles } from "@emotion/core";
-import { ClassNameMap } from "@material-ui/core/styles/withStyles";
 import TableCell from "@material-ui/core/TableCell";
 import {
-    defaultTableBorder,
-    fixedCellSize,
     highlightedTableBorder,
+    fixedCellSize,
+    defaultTableBorder,
     sumRowColBackgroundColor,
-} from "../ResultsTable.style";
+} from "./ResultsTable.style";
 
 const tableCellStyle = (isName: boolean): SerializedStyles => css`
     box-sizing: border-box;
@@ -26,12 +25,12 @@ const totalSumStyle = css`
     background-color: ${sumRowColBackgroundColor};
 `;
 
-export function createTableRowWithColSumService(props: {
+export function TableContentRowWithColSumComponent(props: {
     showRowSum: boolean;
     rowWithColSums: Record<string, string> | undefined;
     headerValues: string[];
+    classes: Record<"tableCell", string>;
     colSumLabel: string;
-    classes: ClassNameMap<"tableCell">;
 }): JSX.Element[] {
     const arrayWithColSumCells: JSX.Element[] = [
         <TableCell
