+++ src/app/Main/Body/QueryPage/Services/ExportServices/StatExportServices/generateStatisticTableCsvString.service.ts
@@ -21,21 +21,12 @@ export function generateStatisticTableCsvString(
 
     if (tableAttributeNames.row !== undefined) {
         const headerToPrint: string[] = [...headers];
-        const index = headerToPrint.indexOf("name");
-        if (index > -1) {
-            headerToPrint.splice(index, 1);
-        }
-        headerToPrint.unshift(tableAttributeNames.row)
+        headerToPrint[0] = tableAttributeNames.row;
         StatDataString.push(headerToPrint.join(","));
     } else {
-        const headerWithoutRowName: string[]= [...headers];
-        const index = headerWithoutRowName.indexOf("name");
-        if (index > -1) {
-            headerWithoutRowName.splice(index, 1);
-        }
-        StatDataString.push(`,${headerWithoutRowName.join(",")}`);
+        StatDataString.push(`,${headers.slice(1).join(",")}`);
     }
-    
+
     StatDataString.push(
         generateStatisticRowsCsvString(statDataSet.statData, headers)
     );
