+++ src/app/Shared/Context/DataContext.tsx
@@ -10,8 +10,6 @@ export interface DataInterface {
     option: DisplayOptionType;
     statisticDataAbsolute: Record<string, string>[];
     statisticDataRelative: Record<string, string>[];
-    statTableDataWithSumsAbs: Record<string, string>[];
-    statTableDataWithSumsRel: Record<string, string>[];
 }
 
 export const defaultData: DataInterface = {
@@ -20,8 +18,6 @@ export const defaultData: DataInterface = {
     option: "absolute",
     statisticDataAbsolute: [],
     statisticDataRelative: [],
-    statTableDataWithSumsAbs: [],
-    statTableDataWithSumsRel: [],
 };
 
 interface ProfileState {
